third_party/blink/web_tests/platform/generic/editing/pasteboard/copy-backslash-with-euc-expected.txt

Original texts which will be copied
\ from UTF8 page
\ FROM UTF8 PAGE (TRANSFORMED)
•••••••••••••••••••••••••
Texts in the EUC frame will be pasted here
from EUC div to UTF8 text-control: 
from EUC div to UTF8 content-editable: \ from EUC page
from EUC transform to UTF8 text-control: 
from EUC transform to UTF8 content-editable: \ FROM EUC PAGE (TRANSFORMED)
from EUC secure to UTF8 text-control: 
from EUC secure to UTF8 content-editable: ••••••••••••••••••••••••
from EUC control to UTF8 text-control: 
from EUC control to UTF8 content-editable: \ from EUC text control



Results

from EUC div to UTF8 text-control: PASS
from EUC div to UTF8 content-editable: PASS
from EUC transform to UTF8 text-control: PASS
from EUC transform to UTF8 content-editable: PASS
from EUC secure to UTF8 text-control: PASS
from EUC secure to UTF8 content-editable: PASS
from EUC control to UTF8 text-control: PASS
from EUC control to UTF8 content-editable: PASS
from UTF8 div to EUC text-control: PASS
from UTF8 div to EUC content-editable: PASS
from UTF8 transform to EUC text-control: PASS
from UTF8 transform to EUC content-editable: PASS
from UTF8 secure to EUC text-control: PASS
from UTF8 secure to EUC content-editable: PASS
from UTF8 control to EUC text-control: PASS
from UTF8 control to EUC content-editable: PASS
